loads ledger from address url string
