+++ i18n/en.i18n.json
@@ -829,5 +829,5 @@
   "guild-total-value": "Total Value",
   "guild-voting-address": "Voter",
   "guild-voting-addresses": "Voters",
-  "replica-person": "Address."
+  "replica-person": "Profile."
 }
