+++ imports/ui/templates/components/identity/login/profile/profileEditor.html
@@ -77,12 +77,3 @@
     </div>
   </div>
 </template>
-
-<template name="qr">
-  <div id="qr-code">
-    <canvas class="qr-code-address" id="canvas-qr"></canvas>
-    <div title="{{_ 'click-to-copy'}}" class="qr-code-address-text">
-      {{_ code}}
-    </div>
-  </div>
-</template>
