+++ imports/ui/templates/widgets/qr/qr.html
@@ -1,0 +1,5 @@
+<template name="qr">
+  <div title="{{_ 'copy-clipboard'}}" id="qr-code">
+    <canvas class="qr-code-address" id="canvas-qr"></canvas>
+  </div>
+</template>
