+++ lib/views.js
@@ -467,7 +467,7 @@ _views.transactionsGeo = (terms) => {
 */
 _views.transactionsPeer = (terms) => {
   let contractId;
-  console.log('----')
+  console.log('----');
   console.log(terms);
   if (!terms.userId && terms.username) {
     contractId = Meteor.users.findOne({ username: terms.username.toLowerCase() })._id;
